Break dependency between language testing and headless workspace
